Added an implementation of postfix using stacks
This implementation of postfix uses linked lists
for stacks as opposed to arrays. So far, the implemtantion
works fine with the four basic operators. Thus, addition,
subtraction, division, and multiplication. To the best of my
knowledge, it should handle postfix arithmetic well. But I'm
consistently learning to see if I can make this better.